feat(app): allow pinning system-software versions on app create
Add a `--set <dependency>=<version>` flag to the custom runtime create
commands (`app create php|php-worker|node|python`) that pins the
system-software (runtime) version at creation time, mirroring
`app dependency update`. Previously the newest available runtime was
always assigned, requiring a second `app dependency update` call.

The dependency-resolution logic is extracted into a shared
`updateAppDependencies` helper used by both the create flow and the
`app dependency update` command.

Closes #1951
