+++ src/commands/app/create/php-worker.tsx
@@ -8,7 +8,7 @@ import {
 export const phpWorkerInstaller = new AppInstaller(
   "fcac178a-e606-4460-a5fd-b3ad0ae7a3cc",
   "PHP worker",
-  ["entrypoint", "site-title"] as const,
+  ["entrypoint", "site-title", "set"] as const,
 );
 
 export default class InstallPhpWorker extends ExecRenderBaseCommand<
