+++ src/commands/app/create/php.tsx
@@ -8,7 +8,7 @@ import {
 export const phpInstaller = new AppInstaller(
   "34220303-cb87-4592-8a95-2eb20a97b2ac",
   "custom PHP",
-  ["document-root", "site-title"] as const,
+  ["document-root", "site-title", "set"] as const,
 );
 
 export default class InstallPhp extends ExecRenderBaseCommand<
@@ -17,6 +17,13 @@ export default class InstallPhp extends ExecRenderBaseCommand<
 > {
   static description = phpInstaller.description;
   static flags = phpInstaller.flags;
+  static examples = [
+    {
+      description: "Create a PHP app pinned to the latest PHP 8.3 release",
+      command:
+        "<%= config.bin %> <%= command.id %> --document-root /public --set php=~8.3",
+    },
+  ];
 
   protected async exec() {
     return phpInstaller.exec(
