+++ src/commands/app/create/python.tsx
@@ -8,7 +8,7 @@ import {
 export const pythonInstaller = new AppInstaller(
   "be57d166-dae9-4480-bae2-da3f3c6f0a2e",
   "custom python site",
-  ["site-title", "entrypoint"] as const,
+  ["site-title", "entrypoint", "set"] as const,
 );
 
 export default class InstallPython extends ExecRenderBaseCommand<
@@ -17,6 +17,12 @@ export default class InstallPython extends ExecRenderBaseCommand<
 > {
   static description = pythonInstaller.description;
   static flags = pythonInstaller.flags;
+  static examples = [
+    {
+      description: "Create a Python app pinned to the latest 3.11 release",
+      command: "<%= config.bin %> <%= command.id %> --set python=~3.11",
+    },
+  ];
 
   protected async exec() {
     return pythonInstaller.exec(
