+++ src/commands/app/dependency/update.tsx
@@ -5,19 +5,13 @@ import {
   processFlags,
 } from "../../../rendering/process/process_flags.js";
 import { Flags } from "@oclif/core";
-import { assertStatus } from "@mittwald/api-client-commons";
 import { ReactNode } from "react";
 import type { MittwaldAPIV2 } from "@mittwald/api-client";
 import { Success } from "../../../rendering/react/components/Success.js";
-import { Range } from "semver";
-import { ProcessRenderer } from "../../../rendering/process/process.js";
-import { compareVersionsBy } from "../../../lib/resources/app/versions.js";
+import { updateAppDependencies } from "../../../lib/resources/app/dependencies.js";
 
 type AppSystemSoftwareUpdatePolicy =
   MittwaldAPIV2.Components.Schemas.AppSystemSoftwareUpdatePolicy;
-type AppSystemSoftwareVersion =
-  MittwaldAPIV2.Components.Schemas.AppSystemSoftwareVersion;
-type AppSystemSoftware = MittwaldAPIV2.Components.Schemas.AppSystemSoftware;
 
 export default class Update extends ExecRenderBaseCommand<typeof Update, void> {
   static summary = "Update the dependencies of an app";
@@ -56,120 +50,20 @@ export default class Update extends ExecRenderBaseCommand<typeof Update, void> {
       this.flags,
       "Updating app dependencies",
     );
-    const systemSoftwares = await process.runStep(
-      "fetching system softwares",
-      async () => {
-        const r = await this.apiClient.app.listSystemsoftwares();
-        assertStatus(r, 200);
 
-        return r.data;
-      },
+    await updateAppDependencies(
+      this.apiClient,
+      process,
+      appInstallationId,
+      this.flags.set,
+      updatePolicy,
     );
-
-    const versionsToUpdate: {
-      [x: string]: {
-        systemSoftwareVersion: string;
-        updatePolicy: AppSystemSoftwareUpdatePolicy;
-      };
-    } = {};
-
-    for (const s of this.flags.set) {
-      const [software, versionSpec] = s.split("=");
-      const parsedVersionSpec = new Range(versionSpec);
-
-      if (!parsedVersionSpec) {
-        throw new Error(
-          `version spec ${versionSpec} is not a valid semver constraint`,
-        );
-      }
-
-      const systemSoftware = systemSoftwares.find(
-        (s) => s.name.toLowerCase() === software.toLowerCase(),
-      );
-      if (!systemSoftware) {
-        throw new Error(`unknown system software ${software}`);
-      }
-
-      const versions = await this.getVersions(
-        process,
-        systemSoftware,
-        versionSpec,
-      );
-      const version = await process.runStep(
-        `determining version for ${software}`,
-        async () => {
-          const exactMatch = versions.find(
-            (v) => v.externalVersion === versionSpec,
-          );
-          if (exactMatch) {
-            return exactMatch;
-          }
-
-          const recommendedVersion = versions.find((v) => v.recommended);
-          if (recommendedVersion) {
-            return recommendedVersion;
-          }
-
-          if (versions.length === 0) {
-            throw new Error(
-              `no versions found for ${software} and version constraint ${versionSpec}`,
-            );
-          }
-
-          return versions[versions.length - 1];
-        },
-      );
-
-      process.addInfo(
-        `selected ${systemSoftware.name} version: ${version.externalVersion}`,
-      );
-
-      versionsToUpdate[systemSoftware.id] = {
-        systemSoftwareVersion: version.id,
-        updatePolicy,
-      };
-    }
-
-    await process.runStep("updating app dependencies", async () => {
-      const r = await this.apiClient.app.patchAppinstallation({
-        appInstallationId,
-        data: {
-          systemSoftware: versionsToUpdate,
-        },
-      });
-
-      assertStatus(r, 204);
-    });
 
     await process.complete(
       <Success>
         The dependencies of this app were successfully updated!
       </Success>,
     );
-  }
-
-  private async getVersions(
-    p: ProcessRenderer,
-    systemSoftware: AppSystemSoftware,
-    versionRange: string,
-  ): Promise<AppSystemSoftwareVersion[]> {
-    const versions = await p.runStep(
-      `fetching versions for ${systemSoftware.name}`,
-      async () => {
-        const r = await this.apiClient.app.listSystemsoftwareversions({
-          systemSoftwareId: systemSoftware.id,
-          queryParameters: {
-            versionRange,
-          },
-        });
-        assertStatus(r, 200);
-
-        return r.data;
-      },
-    );
-
-    versions.sort(compareVersionsBy("internal"));
-    return versions;
   }
 
   protected render(): ReactNode {
